Use Wiring streams for encoding/decoding Variant as CBOR
